:pencil2: feat: 정복자 선정 기준을 추가로 세우고, 반환값으로 자격 여부와 총점을 반환하도록 수정

See merge request s09-final/S09P31A705!140

+++ backend/game-service/src/main/java/com/popplar/gameservice/dto/GameDto.java
@@ -21,13 +21,13 @@ public class GameDto {
     private Long memberId;
 
     @NotNull
-    private int points;
+    private double points;
 
     @NotNull
     private LocalDateTime createdDate;
 
     @Builder
-    public GameDto(GameType type, Long hotPlaceId, int points,
+    public GameDto(GameType type, Long hotPlaceId, double points,
         LocalDateTime createdDate) {
         this.type = type;
         this.memberId=0L;
